requirements.txt
tensorflow==2.19.1
mediapipe==0.10.21
opencv-python==4.11.0.86
numpy==1.26.4
pyttsx3==2.99
supabase==2.27.0
qrcode==8.2
Pillow==12.0.0